// Foodoz_Admin/foodoz-angular/src/app/@core/models/product/product.error.ts
export class ProductError {
    name_translations: Array<string>;
    title_translations: Array<string>;
    detail_translations: Array<string>;
    meta: Array<string>;
    price: Array<string>;
    stock_quantity: Array<string>;
    vendor_id: Array<string>;
    images: Array<string>;
    user_id: Array<string>;
    categories: Array<string>;
}